fix warnings

+++ lib/shapoco/jjy/rx/common.hpp
@@ -5,6 +5,7 @@
 #endif
 
 #include <stdint.h>
+#include <string.h>
 
 namespace shapoco::jjy::rx {
 
@@ -13,7 +14,7 @@ static constexpr uint32_t DETECTION_INPUT_SPS = SHAPOCO_JJY_CFG_RX_ADC_SPS;
 static constexpr uint32_t DETECTION_RESOLUTION = SHAPOCO_JJY_CFG_RX_ADC_SPS / FREQ_60KHZ;
 
 static constexpr uint32_t DETECTION_OUTPUT_SPS = 100;
-static constexpr uint32_t DETECTION_BLOCK_SIZE = DETECTION_INPUT_SPS / DETECTION_OUTPUT_SPS;
+static constexpr size_t DETECTION_BLOCK_SIZE = DETECTION_INPUT_SPS / DETECTION_OUTPUT_SPS;
 
 static_assert(DETECTION_RESOLUTION * FREQ_60KHZ == DETECTION_INPUT_SPS);
 static_assert(DETECTION_BLOCK_SIZE * DETECTION_OUTPUT_SPS == DETECTION_INPUT_SPS);
